Function combo::combinations

pub fn combinations<'a, T>(seq: &'a [T], len: usize) -> Combinator<'a, T>

Returns a (non-standard) iterator yielding all combinations of length len from the sequence seq.

Each combination is itself a (standard) Iterator yielding references to each element of the combination.

Panics

When attempting to iterate over combination lengths longer than the original sequence.

Examples

use combo::combinations;

let sequence: Vec<u32> = (0..5).collect();
let mut combinator = combinations(&sequence[..], 3);

let mut i = 0;
while let Some(combo) = combinator.next() {
    let combination: Vec<&u32> = combo.collect();
    println!("{}: {:?}", i, combination);
    i += 1;
}

// Prints:
// 0: [0, 1, 2]
// 1: [0, 1, 3]
// 2: [0, 1, 4]
// 3: [0, 2, 3]
// 4: [0, 2, 4]
// 5: [0, 3, 4]
// 6: [1, 2, 3]
// 7: [1, 2, 4]
// 8: [1, 3, 4]
// 9: [2, 3, 4]